+++ src/main/java/de/superrental/carrentalinterface/dto/CustomerDTO.java
@@ -4,6 +4,6 @@ public record CustomerDTO(
         String firstName,
         String lastName,
         String email,
-        Long customerId
+        Integer customerId
 )
 {}
